Collect the oblique-shock pilot feature into an _euler package
Move the oblique-shock pilot code from solvcon/pilot/_oblique.py into a new
solvcon.pilot._euler package and register it in setup.py.  Only the mesh
feature moves; the work-in-progress solution menu item and its driver are
dropped, since a dedicated solver panel replaces them.
